MediBot/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_groq
python-dotenv
gunicorn